Add temporary script for generating static storage var identifiers
Starknet requires storage variable accessors to use a computed _hash_
rather than a standard ascii string of the variable name.

The starknet.js library doesn't support generation of storage variable
identifiers yet, so it is necessary to compute and copy these
into source code manually when introducing new storage variables.

In the future it probably makes sense to implement this into the library.