Improve canvas update method to allow patch of multiple pixel changes
These changes improve on the work-in-progress canvas update method in
the smart contract. With these changes, a user can paint a small
segment of the canvas and commit it to the blockchain with a single
transaction.

Eventually this will cost ERC20 $PAINT tokens and reward users for
their artistry. Some rate limiting will need to be handled to avoid
token farming.