Add paint and erase capability to Canvas
These changes add the minimal required mechanics for the frontend.

In the future the current canvas state will need to be painted on load,
as well as on-demand whenever a new block contains updated canvas state.

Additional changes will be required to map color identifiers to color
values; otherwise perhaps a traditional 256^6 hexadecimal color code
will be stored if the storage on Starknet is cheap enough.