Add ERC20 $PAINT token and pay users for updates to canvas
These changes introduce the basis for the $PAINT ERC20 token and
play-to-earn gameplay tokenomics. Users will now be rewarded one
$PAINT token per pixel that is updated.

In the future each update will have an associated cost. Payout and costs
will scale as the size and update frequency of the canvas increases;
however, exact details on the tokenomics are TBD.