Implement restricted user activity by time duration
These changes use the player last activity timestamps to restrict
gameplay with time duration throttling.

Without this restrictions it is possible for a script to overtake the
canvas. It introduces an undiscovered gameplay mechanic where the amount
of $PAINT held correlates with the player's ability to paint and control
the canvas.