Add required $PAINT cost per updated pixel
These changes add an update cost to every canvas paint. This restricts
new users from dominating the canvas.

Required $PAINT is not deducted from the user. It grants permission for
an allowance. It does not get consumed as a utility token.

Eventually there will be user actions which consume $PAINT. Balance must
be introduced so players do not dominate the canvas.